trader/requirements.txt
httpx==0.27.2
pydantic==2.8.2
numpy>=2.2.0
pandas>=2.2.2
python-dotenv==1.0.1
orjson==3.10.7
tenacity==9.0.0
SQLAlchemy[asyncio]>=2.0.35
asyncpg>=0.30.0
